+++ lib/muffin_man/vendor_direct_fulfillment_payments/v1.rb
@@ -1,0 +1,12 @@
+module MuffinMan
+  module VendorDirectFulfillmentPayments
+    class V1 < SpApiClient
+      def submit_invoice(invoices)
+        @local_var_path = "/vendor/directFulfillment/payments/v1/invoices"
+        @request_body = { "invoices" => invoices }
+        @request_type = "POST"
+        call_api
+      end
+    end
+  end
+end