+++ lib/muffin_man/vendor_transaction_status/v1.rb
@@ -1,0 +1,11 @@
+module MuffinMan
+  module VendorTransactionStatus
+    class V1 < SpApiClient
+      def get_transaction(transaction_id)
+        @local_var_path = "/vendor/transactions/v1/transactions/#{transaction_id}"
+        @request_type = "GET"
+        call_api
+      end
+    end
+  end
+end